moved routes to controllers folder
